Final project

+++ MyCollege.Data/Models/Grade.cs
@@ -12,7 +12,7 @@ namespace MyCollege.Data.Models
         public int Id { get; set; }
         public int StudentId { get; set; }
         public int SubjectId { get; set; }
-        public decimal Value { get; set; }
+        public int Value { get; set; }
 
         public virtual Student Student { get; set; }
         public virtual Subject Subject { get; set; }
